src/requirements.txt
pandas
numpy
matplotlib
pyyaml
datacommon[dbconnectors]
